Pass helpers to params and prompt hooks
