+++ src/__tests__/metaverse/hygen-templates/_templates/inflection-in-params/new/index.ejs.t
@@ -1,0 +1,4 @@
+---
+to: given/inflection-in-params/new.md
+---
+<%= pluralizedName %>
