+++ src/__tests__/metaverse/hygen-templates/_templates/inflection-in-params/new/index.js
@@ -1,0 +1,5 @@
+module.exports = {
+  params: ({ args, h }) => {
+    return { pluralizedName: h.inflection.pluralize(args.name) }
+  },
+}
