+++ src/__tests__/metaverse/hygen-templates/expected/inflection-in-params/new.md
@@ -1,0 +1,1 @@
+people
