+++ src/context.ts
@@ -1,26 +1,9 @@
 import { RunnerConfig } from './types'
-import inflection from 'inflection'
-import changeCase from 'change-case'
-import path from 'path'
+import helpers from './helpers'
 
 const localsToCapitalize = ['name']
 const localsDefaults = {
   name: 'unnamed',
-}
-// supports kebab-case to KebabCase
-inflection.undasherize = (str) =>
-  str
-    .split(/[-_]/)
-    .map((w) => w[0].toUpperCase() + w.slice(1).toLowerCase())
-    .join('')
-
-const helpers = {
-  capitalize(string) {
-    return string.charAt(0).toUpperCase() + string.slice(1)
-  },
-  inflection,
-  changeCase,
-  path,
 }
 
 const doCapitalization = (hsh, [key, value]) => {
