+++ src/prompt.ts
@@ -1,6 +1,7 @@
 import path from 'path'
 import fs from 'fs'
 import { Prompter } from './types'
+import helpers from './helpers'
 
 const hooksfiles = ['prompt.js', 'index.js']
 const prompt = <Q, T>(
@@ -9,8 +10,8 @@ const prompt = <Q, T>(
   args: Record<string, any>,
 ): Promise<T | object> => {
   const hooksfile = hooksfiles
-    .map(f => path.resolve(path.join(actionfolder, f)))
-    .find(f => fs.existsSync(f))
+    .map((f) => path.resolve(path.join(actionfolder, f)))
+    .find((f) => fs.existsSync(f))
 
   if (!hooksfile) {
     return Promise.resolve({})
@@ -20,20 +21,25 @@ const prompt = <Q, T>(
   // $FlowFixMe
   const hooksModule = require(hooksfile)
   if (hooksModule.params) {
-    return hooksModule.params({ args })
+    return hooksModule.params({ args, h: helpers })
   }
 
   // lazy loads prompter
   // everything below requires it
   const prompter = createPrompter()
   if (hooksModule.prompt) {
-    return hooksModule.prompt({ prompter, inquirer: prompter, args })
+    return hooksModule.prompt({
+      prompter,
+      inquirer: prompter,
+      args,
+      h: helpers,
+    })
   }
 
   return prompter.prompt(
     // prompt _only_ for things we've not seen on the CLI
     hooksModule.filter(
-      p =>
+      (p) =>
         args[p.name] === undefined ||
         args[p.name] === null ||
         args[p.name].length === 0,
